Fixed the submit button on qualify page, moved logic from app.js to qualify.js
