+++ routes/qualify.js
@@ -1,14 +1,14 @@
 var express = require('express');
 var router = express.Router();
+//var lodash = require('lodash');
 // var app = express();
 
 /* GET headers page. */
 
 router.get('/', function(req, res, next) {
+	console.log("*************");
     res.render('qualify', { title: 'Express' });
 });
-
-
 
 
 module.exports = router;
