+++ views/decision.ejs
@@ -1,0 +1,40 @@
+
+<% include templates/header.ejs %>
+<main>
+    <article>
+        <!-- Header -->
+        <header>
+            <div class="container">
+                <div class="row">
+                    <div class="col-sm-12">
+                        <div class="intro-text">
+                            <span class="name">Affordable Housing Program</span>
+                             <span class="skills">Princeton, New Jersey</span>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </header>
+        <section>
+            <div class="row">
+                <div class="col-xs-12">
+                    <h2>Check eligibility to join the program.</h2>
+                </div>
+            </div>
+            <div class="row">
+                <div class="col-xs-12 col-sm-7 pull-left">
+                    <br/><br/><br/>
+                    <p><%- message %></p> 
+                </div>
+                <div class="col-xs-12 col-sm-5 pull-right">
+                    <figure>
+                        <img class="img img-responsive side-image" src="http://placehold.it/300x300">
+                    </figure>
+                </div>
+                
+                
+                </div>
+            </div>
+        </section>
+    </main>
+  <% include templates/footer.ejs %>
