+++ views/qualify.ejs
@@ -28,7 +28,7 @@
                 </div>
                 <div class="col-xs-12 col-sm-7 pull-left">
                     <p>Fill the below info to check if you are qualified to join the program!</p>
-                    <form action="" method="POST">
+                    <form action="/decision" method="POST">
                         <dl>
                             <dt><label for="first_name">First Name</label></dt>
                             <dd><input type="text" id="first_name" name="first_name" /></dd>
@@ -39,7 +39,7 @@
                             <dt><label for="total_salary">Total Annual Salary for the household</label></dt>
                             <dd><input type="number" id="total_salary" name="total_salary" /></dd>
                         </dl>
-                        <input class="btn btn-princeton" type="submit" value="Check My Eligibility">
+                        <input class="btn btn-princeton" href="./decision" type="submit" value="Check My Eligibility">
                     </form>
                 </div>
             </div>
